Alteração na Questão 3: valores alterados para Float

+++ Questão 3/app.js
@@ -23,7 +23,7 @@ while(aux<=3){
 }
 
     else{
-        soma=((parseInt(soma))+parseInt((notas[aux])));
+        soma=((parseFloat(soma))+parseFloat((notas[aux])));
         aux++;
     }
 }
